Renamed Interval.singleton to Interval.discrete
