+++ plutus-ledger-api/src/PlutusLedgerApi/Common/Eval.hs
@@ -1,147 +1,9 @@
--- editorconfig-checker-disable-file
-{-# LANGUAGE OverloadedStrings #-}
-{-# LANGUAGE StrictData        #-}
-{-# LANGUAGE TemplateHaskell   #-}
-{-# LANGUAGE TypeApplications  #-}
-{-# OPTIONS_GHC -Wno-deferred-out-of-scope-variables #-}
 module PlutusLedgerApi.Common.Eval
-    ( EvaluationError (..)
-    , EvaluationContext
-    , VerboseMode (..)
-    , LogOutput
-    , evaluateScriptCounting
-    , evaluateScriptRestricting
+    ( module PlutusLedgerApi.Common.Versions
+    , module PlutusLedgerApi.Internal.Eval
     ) where
 
-import PlutusCore
-import PlutusCore as ScriptPlutus (Version, defaultVersion)
-import PlutusCore.Data as Plutus
-import PlutusCore.Evaluation.Machine.ExBudget as Plutus
-import PlutusCore.MkPlc qualified as UPLC
-import PlutusCore.Pretty
-import PlutusLedgerApi.Common.SerialisedScript
 import PlutusLedgerApi.Common.Versions
-import PlutusPrelude
-import UntypedPlutusCore qualified as UPLC
-import UntypedPlutusCore.Evaluation.Machine.Cek qualified as UPLC
-import PlutusLedgerApi.Internal.EvaluationContext
-import PlutusLedgerApi.Internal.SerialisedScript
-
-import Control.Monad.Except
-import Control.Monad.Writer
-import Data.Text as Text
-import Data.Tuple
-import Prettyprinter
-import Control.Lens
-
--- | Errors that can be thrown when evaluating a Plutus script.
-data EvaluationError =
-    CekError (UPLC.CekEvaluationException NamedDeBruijn DefaultUni DefaultFun) -- ^ An error from the evaluator itself
-    | DeBruijnError FreeVariableError -- ^ An error in the pre-evaluation step of converting from de-Bruijn indices
-    | CodecError ScriptDecodeError -- ^ A deserialisation error
-    | IncompatibleVersionError (ScriptPlutus.Version ()) -- ^ An error indicating a version tag that we don't support
-    -- TODO: make this error more informative when we have more information about what went wrong
-    | CostModelParameterMismatch -- ^ An error indicating that the cost model parameters didn't match what we expected
-    deriving stock (Show, Eq)
-makeClassyPrisms ''EvaluationError
-
-instance AsScriptDecodeError EvaluationError where
-    _ScriptDecodeError = _CodecError
-
-instance Pretty EvaluationError where
-    pretty (CekError e)      = prettyClassicDef e
-    pretty (DeBruijnError e) = pretty e
-    pretty (CodecError e) = viaShow e
-    pretty (IncompatibleVersionError actual) = "This version of the Plutus Core interface does not support the version indicated by the AST:" <+> pretty actual
-    pretty CostModelParameterMismatch = "Cost model parameters were not as we expected"
-
--- | The type of log output: just a list of 'Text'.
---
--- TODO: api-user does not depend on it, so move it to plutus-core and also perhaps rename it
-type LogOutput = [Text.Text]
-
--- | A simple toggle indicating whether or not we should produce logs.
-data VerboseMode = Verbose | Quiet
-    deriving stock (Eq)
-
-{-|
-Evaluates a script, with a cost model and a budget that restricts how many
-resources it can use according to the cost model. Also returns the budget that
-was actually used.
-
-Can be used to calculate budgets for scripts, but even in this case you must give
-a limit to guard against scripts that run for a long time or loop.
-
-Note: Parameterized over the ledger-plutus-version since the builtins allowed (during decoding) differs.
--}
-evaluateScriptRestricting
-    :: PlutusVersion
-    -> ProtocolVersion
-    -> VerboseMode     -- ^ Whether to produce log output
-    -> EvaluationContext -- ^ The cost model that should already be synced to the most recent cost-model-params coming from the current protocol
-    -> ExBudget        -- ^ The resource budget which must not be exceeded during evaluation
-    -> SerialisedScript          -- ^ The script to evaluate
-    -> [Plutus.Data]          -- ^ The arguments to the script
-    -> (LogOutput, Either EvaluationError ExBudget)
-evaluateScriptRestricting lv pv verbose ectx budget p args = swap $ runWriter @LogOutput $ runExceptT $ do
-    appliedTerm <- mkTermToEvaluate lv pv p args
-
-    let (res, UPLC.RestrictingSt (ExRestrictingBudget final), logs) =
-            UPLC.runCekDeBruijn
-                (toMachineParameters pv ectx)
-                (UPLC.restricting $ ExRestrictingBudget budget)
-                (if verbose == Verbose then UPLC.logEmitter else UPLC.noEmitter)
-                appliedTerm
-
-    tell logs
-    liftEither $ first CekError $ void res
-    pure (budget `minusExBudget` final)
-
-{-|
-Evaluates a script, returning the minimum budget that the script would need
-to evaluate successfully. This will take as long as the script takes, if you need to
-limit the execution time of the script also, you can use 'evaluateScriptRestricting', which
-also returns the used budget.
-
-Note: Parameterized over the ledger-plutus-version since the builtins allowed (during decoding) differs.
--}
-evaluateScriptCounting
-    :: PlutusVersion
-    -> ProtocolVersion
-    -> VerboseMode     -- ^ Whether to produce log output
-    -> EvaluationContext -- ^ The cost model that should already be synced to the most recent cost-model-params coming from the current protocol
-    -> SerialisedScript          -- ^ The script to evaluate
-    -> [Plutus.Data]          -- ^ The arguments to the script
-    -> (LogOutput, Either EvaluationError ExBudget)
-evaluateScriptCounting lv pv verbose ectx p args = swap $ runWriter @LogOutput $ runExceptT $ do
-    appliedTerm <- mkTermToEvaluate lv pv p args
-
-    let (res, UPLC.CountingSt final, logs) =
-            UPLC.runCekDeBruijn
-                (toMachineParameters pv ectx)
-                UPLC.counting
-                (if verbose == Verbose then UPLC.logEmitter else UPLC.noEmitter)
-                appliedTerm
-
-    tell logs
-    liftEither $ first CekError $ void res
-    pure final
-
--- | Shared helper for the evaluation functions, deserialises the 'SerialisedScript' , applies it to its arguments, puts fakenamedebruijns, and scope-checks it.
-mkTermToEvaluate
-    :: (MonadError EvaluationError m)
-    => PlutusVersion
-    -> ProtocolVersion
-    -> SerialisedScript
-    -> [Plutus.Data]
-    -> m (UPLC.Term UPLC.NamedDeBruijn DefaultUni DefaultFun ())
-mkTermToEvaluate lv pv bs args = do
-    -- It decodes the program through the optimized ScriptForExecution. See `ScriptForExecution`.
-    ScriptForExecution (UPLC.Program _ v t) <- deserialiseScriptForExecution lv pv bs
-    unless (v == ScriptPlutus.defaultVersion ()) $ throwError $ IncompatibleVersionError v
-    let termArgs = fmap (UPLC.mkConstant ()) args
-        appliedT = UPLC.mkIterApp () t termArgs
-
-    -- make sure that term is closed, i.e. well-scoped
-    through (liftEither . first DeBruijnError . UPLC.checkScope) appliedT
-
+import PlutusLedgerApi.Internal.Eval
+    hiding ( mkDynEvaluationContext
+           )
