+++ plutus-ledger-api/src/PlutusLedgerApi/V1/ParamName.hs
@@ -1,7 +1,7 @@
 {-# LANGUAGE DerivingVia #-}
 module PlutusLedgerApi.V1.ParamName
     ( ParamName
-    , module PlutusLedgerApi.Internal.IsParamName
+    , IsParamName (showParamName)
     ) where
 
 import Data.Ix
