+++ plutus-ledger-api/src/PlutusLedgerApi/V2/Tx.hs
@@ -41,12 +41,14 @@ import PlutusTx qualified
 import PlutusTx.Bool qualified as PlutusTx
 import PlutusTx.Eq qualified as PlutusTx
 
-import PlutusLedgerApi.V1.Address
-import PlutusLedgerApi.V1.Crypto
-import PlutusLedgerApi.V1.Scripts
 import PlutusLedgerApi.V1.Tx hiding (TxOut (..), isPayToScriptOut, isPubKeyOut, outAddress, outValue, pubKeyHashTxOut,
                               txOutDatum, txOutPubKey)
-import PlutusLedgerApi.V1.Value
+
+
+import PlutusLedgerApi.Common.Address
+import PlutusLedgerApi.Common.Crypto
+import PlutusLedgerApi.Common.Scripts
+import PlutusLedgerApi.Common.Value
 
 -- | The datum attached to an output: either nothing; a datum hash; or the datum itself (an "inline datum").
 data OutputDatum = NoOutputDatum | OutputDatumHash DatumHash | OutputDatum Datum
