+++ plutus-ledger-api/src/PlutusLedgerApi/V3/Contexts.hs
@@ -1,0 +1,5 @@
+module PlutusLedgerApi.V3.Contexts
+    ( module PlutusLedgerApi.V2.Contexts
+    ) where
+
+import PlutusLedgerApi.V2.Contexts
