+++ plutus-ledger-api/src/PlutusLedgerApi/V3/Tx.hs
@@ -1,0 +1,5 @@
+module PlutusLedgerApi.V3.Tx
+    ( module PlutusLedgerApi.V2.Tx
+    ) where
+
+import PlutusLedgerApi.V2.Tx
